apps/api/requirements.txt
flask
python-socketio[client]
eventlet 
flask-cors
